backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
ultralytics>=8.3.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
supabase==2.7.0
python-dotenv==1.0.1
pyjwt==2.9.0
pydantic-settings==2.5.2
